Move IPFS file adder and getter to separate node #107
